New json-extension project structure
    Check if extension is started in debug mode
    Fix appveyor.yaml
    Update versions
    Update READMEs
