+++ examples/README.md
@@ -1,14 +1,3 @@
 # pygls examples
 
-## Json Language Server
-
-### Run Server
-
-1. Open root directory in VS Code
-1. Open debug view (`ctrl + shift + D`).
-1. Select `Launch Json-Extension server` and press `F5`.
-
-### Run Client
-
-1. Open `vscode-client` and run `npm install`.
-1. Select `Launch Json-Extension client` and press `F5`.
+- [Json Language Server](https://github.com/openlawlibrary/pygls/tree/master/examples/json-extension)
