+++ examples/json-extension/.gitignore
@@ -1,0 +1,5 @@
+out
+node_modules
+client/server
+.vscode-test
+.vscode/settings.json
