+++ examples/json-extension/.vscode/launch.json
@@ -1,36 +1,39 @@
+// A launch configuration that compiles the extension and then opens it inside a new window
 {
-    "version": "0.2.0",
-    "configurations": [
-        {
-            "name": "Launch JsonExtension Client",
-            "type": "extensionHost",
-            "request": "launch",
-            "runtimeExecutable": "${execPath}",
-            "args": [
-                "--extensionDevelopmentPath=${workspaceRoot}/vscode-client"
-            ],
-            "stopOnEntry": false,
-            "sourceMaps": true,
-            "outFiles": [
-                "${workspaceRoot}/vscode-client/out/src/**/*.js"
-            ],
-            "preLaunchTask": "json-extension-watch:client"
-        },
-        {
-            "name": "Launch JsonExtension Server",
-            "type": "python",
-            "request": "launch",
-            "console": "none",
-            "module": "server",
-            "args": [
-                "--tcp"
-            ],
-            "debugStdLib": true,
-            "pythonPath": "${config:python.pythonPath}",
-            "cwd": "${workspaceFolder}",
-            "env": {
-                "PYTHONPATH": "${workspaceFolder}"
-            }
-        }
-    ]
+  "version": "0.2.0",
+  "configurations": [
+    {
+      "type": "extensionHost",
+      "request": "launch",
+      "name": "Launch Client",
+      "runtimeExecutable": "${execPath}",
+      "sourceMaps": true,
+      "args": ["--extensionDevelopmentPath=${workspaceRoot}"],
+      "outFiles": ["${workspaceRoot}/client/out/**/*.js"],
+      "preLaunchTask": "json-extension-watch:client",
+      "env": {
+        "VSCODE_DEBUG_MODE": "true"
+      }
+    },
+    {
+      "name": "Launch Server",
+      "type": "python",
+      "request": "launch",
+      "console": "none",
+      "module": "server",
+      "args": ["--tcp"],
+      "debugStdLib": true,
+      "pythonPath": "${config:python.pythonPath}",
+      "cwd": "${workspaceFolder}",
+      "env": {
+        "PYTHONPATH": "${workspaceFolder}"
+      }
+    }
+  ],
+  "compounds": [
+    {
+      "name": "Server + Client",
+      "configurations": ["Launch Server", "Launch Client"]
+    }
+  ]
 }
