--- examples/json-extension/.vscode/settings.json
@@ -1,4 +1,0 @@
-{
-    "python.pythonPath": "/Users/daniel/.virtualenvs/pygls/bin/python",
-    "restructuredtext.confPath": "${workspaceFolder}/docs/source"
-}
