+++ examples/json-extension/.vscodeignore
@@ -1,0 +1,13 @@
+.vscode
+.gitignore
+
+client
+
+tsconfig.json
+tslint.json
+
+out/*.map
+
+package.json
+package-lock.json
+
