+++ examples/json-extension/client/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "json-extension-client",
+  "description": "Simple json extension example",
+  "author": "Open Law Library",
+  "license": "All rights reserved",
+  "version": "0.6.0",
+  "publisher": "openlawlibrary",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/openlawlibrary/pygls"
+  },
+  "engines": {
+    "vscode": "^1.30.1"
+  },
+  "scripts": {
+    "update-vscode": "vscode-install",
+    "postinstall": "vscode-install"
+  },
+  "dependencies": {
+    "vscode-languageclient": "^5.2.1"
+  },
+  "devDependencies": {
+    "vscode": "^1.1.26"
+  }
+}
