+++ examples/json-extension/package.json
@@ -7,7 +7,7 @@
   "version": "0.6.0",
   "publisher": "openlawlibrary",
   "engines": {
-    "vscode": "^1.29.1"
+    "vscode": "^1.30.1"
   },
   "categories": [
     "Other"
@@ -60,13 +60,13 @@
   },
   "devDependencies": {
     "@types/node": "^8.9.4",
-    "tslint": "^5.11.0",
-    "tslint-microsoft-contrib": "^5.2.1",
-    "tsutils": "^3.5.1",
-    "typescript": "^3.2.1"
+    "tslint": "^5.12.0",
+    "tslint-microsoft-contrib": "^6.0.0",
+    "tsutils": "^3.5.2",
+    "typescript": "^3.2.2"
   },
   "dependencies": {
-    "vscode": "^1.1.22",
-    "vscode-languageclient": "^5.1.1"
+    "vscode": "^1.1.26",
+    "vscode-languageclient": "^5.2.1"
   }
 }
