--- examples/json-extension/vscode-client/.gitignore
@@ -1,4 +1,0 @@
-out
-node_modules
-.vscode-dev
-/server
