--- examples/json-extension/vscode-client/.vscodeignore
@@ -1,9 +1,0 @@
-.vscode/**
-typings/**
-out/test/**
-test/**
-src/**
-**/*.map
-.gitignore
-tsconfig.json
-vsc-extension-quickstart.md
