--- examples/json-extension/vscode-client/ThirdPartyNotices.txt
@@ -1,1 +1,0 @@
-Please see ThirdPartyNotices.txt (https://github.com/openlawlibrary/pygls/blob/master/ThirdPartyNotices.txt) at the project root.
